requirements.txt
requests
pyserial
indipy>=0.3.0
django-environ
